Crate problem037

Problem 37 (Truncatable primes)

Solution detail

There is an easy method to generate all the right-truncatable primes: start with the single digit primes 2, 3, 5 and 7 as roots, and do a depth-first traversal of the tree of numbers formed by adding one of the digits 1, 3, 7, 9 to the end of the previous number. We are only interested in the nodes which represent prime numbers, so we prune any branches leading to a non-prime.

Then, to solve the problem, simply iterate over the right-truncatable primes, and check which of them are also left-truncatable.

Note that we have no a priori upper bound on how large these primes can get, so we will perform primality testing using a prime sieve that starts at some small size and grows when necessary.

Structs

TruncatablePrimeTree
The information that is held about the current state during the tree search.

TruncatablePrimeTreeStep
A description of a step that can be taken in the search tree.

Functions

answer
Solve the problem, returning the answer as a String

main

solve
Find the sum of all primes which are both left-truncatable and right-truncatable. Remember to subtract the single-digit primes which are not counted.